+++ src/pages/v2-bug-notice.tsx
